Crate prusti_contracts

Macros

body_invariant

A macro for writing a loop body invariant.

closure

A macro for defining a closure with a specification.

predicate

A macro for defining a predicate using prusti expression syntax instead of just Rust expressions.

Functions

before_expiry

This function is used to evaluate an expression in the context just before the borrows expires.

exists

forall

old

This function is used to evaluate an expression in the “old” context, that is at the beginning of the method call.

Attribute Macros

after_expiry

A macro for writing a pledge on a function.

ensures

A macro for writing a postcondition on a function.

extern_spec

A macro for specifying external functions.

pure

A macro for marking a function as pure.

refine_trait_spec

A macro for impl blocks that refine trait specifications.

requires

A macro for writing a precondition on a function.

trusted

A macro for marking a function as trusted.